First batch of code deletion
