+++ GrailDiary/NotebookStructureViewController.swift
@@ -470,37 +470,15 @@ private extension NotebookStructureViewController {
       bookImporterViewController.delegate = self
       self.present(bookImporterViewController, animated: true)
     }
-    let export = (database as? LegacyNoteDatabase).flatMap { legacyNoteDatabase -> UIAction in
-      UIAction(title: "Export") { [weak self] _ in
-        self?.exportToKVCRDT(legacyNoteDatabase: legacyNoteDatabase)
-      }
-    }
     return UIBarButtonItem(
       image: UIImage(systemName: "ellipsis.circle"),
       menu: UIMenu(
         children: [
           openCommand,
           importLibraryThing,
-          export,
-        ].compactMap { $0 }
+        ]
       )
     )
-  }
-
-  private func exportToKVCRDT(legacyNoteDatabase: LegacyNoteDatabase) {
-    do {
-      let exportedFile = legacyNoteDatabase.fileURL.deletingPathExtension().appendingPathExtension("kvcrdt").lastPathComponent
-      let documentsDirectoryURL = try FileManager.default.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: true)
-      let exportedURL = documentsDirectoryURL.appendingPathComponent(exportedFile)
-      Logger.shared.info("Exporting to \(exportedURL.path)")
-      try legacyNoteDatabase.exportToKVCRDT(exportedURL)
-      let alert = UIAlertController(title: "Export Complete", message: "Export was successful", preferredStyle: .alert)
-      let okAction = UIAlertAction(title: "OK", style: .default)
-      alert.addAction(okAction)
-      present(alert, animated: true)
-    } catch {
-      Logger.shared.error("Unexpected error exporting data: \(error)")
-    }
   }
 
   func updateSnapshot() {
