+++ GrailDiary/Sqlite/NoteDatabase+ImageStorage.swift
@@ -74,90 +74,3 @@ public extension ParsedAttributedString.Style {
     return copy
   }
 }
-
-public extension LegacyNoteDatabase {
-  // TODO: Remove AssetRecord from the schema
-  /// Stores arbitrary data in the database.
-  /// - Parameters:
-  ///   - data: The asset data to store
-  ///   - key: A unique key for the data
-  /// - Throws: .databaseIsNotOpen
-  /// - Returns: The key??
-  @available(*, deprecated, message: "Use writeAssociatedData: instead")
-  func storeAssetData(_ data: Data, key: String) throws -> String {
-    guard let dbQueue = dbQueue else {
-      throw NoteDatabaseError.databaseIsNotOpen
-    }
-    return try dbQueue.write { db in
-      let asset = AssetRecord(id: key, data: data)
-      try asset.save(db)
-      return key
-    }
-  }
-
-  /// Gets arbitrary data back from
-  /// - Parameter key: Key for the asset data to retrieve.
-  /// - Throws: .databaseIsNotOpen, .noSuchAsset
-  /// - Returns: The data corresponding with `key`
-  func retrieveAssetDataForKey(_ key: String) throws -> Data {
-    guard let dbQueue = dbQueue else {
-      throw NoteDatabaseError.databaseIsNotOpen
-    }
-    guard let record = try dbQueue.read({ db in
-      try AssetRecord.filter(key: key).fetchOne(db)
-    }) else {
-      throw NoteDatabaseError.noSuchAsset
-    }
-    return record.data
-  }
-
-  func writeAssociatedData(
-    _ data: Data,
-    noteIdentifier: Note.Identifier,
-    role: String,
-    type: UTType,
-    key: String?
-  ) throws -> String {
-    guard let dbQueue = dbQueue else {
-      throw NoteDatabaseError.databaseIsNotOpen
-    }
-    let actualKey = key ?? ["./" + data.sha1Digest(), type.preferredFilenameExtension].compactMap { $0 }.joined(separator: ".")
-    let binaryRecord = BinaryContentRecord(
-      blob: data,
-      noteId: noteIdentifier,
-      key: actualKey,
-      role: role,
-      mimeType: type.preferredMIMEType ?? "application/octet-stream"
-    )
-    try dbQueue.write { db in
-      try binaryRecord.save(db)
-    }
-    return actualKey
-  }
-
-  func readAssociatedData(from noteIdentifier: Note.Identifier, key: String) throws -> Data {
-    guard let dbQueue = dbQueue else {
-      throw NoteDatabaseError.databaseIsNotOpen
-    }
-    return try dbQueue.read { db in
-      guard let record = try BinaryContentRecord.fetchOne(
-        db,
-        key: [BinaryContentRecord.Columns.noteId.rawValue: noteIdentifier, BinaryContentRecord.Columns.key.rawValue: key]
-      ) else {
-        throw NoteDatabaseError.noSuchAsset
-      }
-      return record.blob
-    }
-  }
-}
-
-// extension NoteDatabase: ImageStorage {
-//  public func storeImageData(_ imageData: Data, suffix: String) throws -> String {
-//    let key = imageData.sha1Digest() + "." + suffix
-//    return try storeAssetData(imageData, key: key)
-//  }
-//
-//  public func retrieveImageDataForKey(_ key: String) throws -> Data {
-//    return try retrieveAssetDataForKey(key)
-//  }
-// }
